Refactor GameDiscovery and enhance feedback tracking
Moved GameDiscovery component to pages directory and updated imports. Enhanced feedback tracking by storing the full Game object in feedbackStore and displaying game details in GameDiscovery. Optimized LikeDislike sentiment state with useMemo.
