+++ src/components/LikeDislike.tsx
@@ -2,6 +2,7 @@ import { Game } from "@/entities/Game";
 import { Sentiment } from "@/entities/Sentiment";
 import { useFeedbackStore } from "@/feedbackStore";
 import { HStack } from "@chakra-ui/react";
+import { useMemo } from "react";
 import { FaHeart, FaHeartBroken } from "react-icons/fa";
 
 interface Props {
@@ -13,11 +14,14 @@ const LikeDislike = ({ game }: Props) => {
   const feedback = useFeedbackStore((s) => s.feedback);
 
   if (import.meta.env.VITE_GAME_DISCOVERY !== "enabled") return null;
-  const state = feedback[game.id]?.sentiment ?? Sentiment.Neutral;
+  const state = useMemo(() => {
+    return feedback[game.id]?.sentiment ?? Sentiment.Neutral;
+  }, [feedback, game.id]);
 
   const toggleSentiment = (target: Sentiment) => {
     setFeedback(game.id, {
       ...feedback[game.id],
+      game,
       sentiment: state === target ? Sentiment.Neutral : target,
     });
   };
