+++ src/components/NavBar.tsx
@@ -4,7 +4,7 @@ import ColorModeSwitch from "./ColorModeSwitch";
 import SearchBar from "./SearchBar";
 import { Link } from "react-router-dom";
 import NavBack from "./NavBack";
-import GameDiscovery from "./GameDiscovery";
+import GameDiscovery from "../pages/GameDiscovery";
 
 const NavBar = () => {
   return (
