+++ src/pages/GameDiscovery.tsx
@@ -1,7 +1,19 @@
-import { Button, CloseButton, Drawer, Portal } from "@chakra-ui/react";
+import { Sentiment } from "@/entities/Sentiment";
+import { useFeedbackStore } from "@/feedbackStore";
+import {
+  Box,
+  Button,
+  CloseButton,
+  Drawer,
+  Portal,
+  VStack,
+  Text,
+} from "@chakra-ui/react";
 import { FaHeartbeat } from "react-icons/fa";
 
 const GameDiscovery = () => {
+  const feedback = useFeedbackStore((s) => s.feedback);
+
   if (import.meta.env.VITE_GAME_DISCOVERY !== "enabled") return null;
   return (
     <Drawer.Root>
@@ -18,10 +30,19 @@ const GameDiscovery = () => {
               <Drawer.Title>Game Discovery</Drawer.Title>
             </Drawer.Header>
             <Drawer.Body>
-              <p>
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do
-                eiusmod tempor incididunt ut labore et dolore magna aliqua.
-              </p>
+              <VStack align="start">
+                {Object.entries(feedback)
+                  .filter(([_, entry]) => entry.sentiment !== Sentiment.Neutral)
+                  .map(([gameId, entry]) => (
+                    <Box key={gameId} p={2}>
+                      {entry.game?.name && (
+                        <Text fontWeight="bold">Name: {entry.game.name}</Text>
+                      )}
+                      <Text>Game ID: {gameId}</Text>
+                      <Text>Sentiment: {entry.sentiment}</Text>
+                    </Box>
+                  ))}
+              </VStack>
             </Drawer.Body>
             <Drawer.CloseTrigger asChild>
               <CloseButton size="md" />
